Refactor register
