+++ src/pages/authentication/index.ts
@@ -1,5 +1,5 @@
 export {default as LoginForm} from './login/login';
-export {default as RegisterForm} from './RegisterForm';
+export {default as RegisterForm} from './register/register';
 export {default as RequestPasswordForm} from './RequestPasswordForm';
 export {default as SetPasswordForm} from './SetPasswordForm';
 export {default as VerifyForm} from './VerifyForm';
